requirements.txt
pandas_datareader
matplotlib.pyplot
pandas
datetime
urllib.request
json
os
numpy
tensorflow
sklearn.preprocessing
